requirements.txt
python
streamlit
google-generativeai
python-dotenv
langchain
langchain-google-genai
tavily-python